refactor(qwen3_5): promote chat_common helpers for cross-model reuse
Move image hashing, stream error signaling, ChatML delta builders, and
the IMAGE_CHANGE_REQUIRES_SESSION_RESTART error prefix from the dense
model into chat_common so MoE, LFM2, Qwen3, Gemma4, and Qianfan-OCR can
share the same session machinery in upcoming steps.
